controllers separated, new customErrors added, and tests aswell
