+++ controllers/chatControllers/createChatController.js
@@ -4,6 +4,7 @@ const User = require('../../models/userModel')
 const { StatusCodes } = require('http-status-codes')
 const getCreatorIdFromHeaders = require('../../middlewares/getCreatorIdFromHeaders')
 const { daysToDieError, usersAlreadyHaveMutualPrivateChatroomError } = require('../../errors/chatErrors/chatErrors')
+const { noUserFoundError } = require('../../errors/userErrors/userErrors')
 
 const createChat = tryCatchWrapper(async (req, res) => {
     let { name: name, is_ttl: isTtl, days_to_die: daysToDie, is_private: isPrivate, usernames: usernames } = req.body
@@ -11,16 +12,13 @@ const createChat = tryCatchWrapper(async (req, res) => {
     for (const username of usernames) {
         let otherUser = await User.findOne({ username: username })
         if (!otherUser) {
-            res.status(StatusCodes.NOT_FOUND).json({
-                message: `No user found with this name: '${usernames}'`,
-            })
+            throw new noUserFoundError(username)
         }
-        if(isPrivate){
+        if (isPrivate) {
             if (await hasMutualPrivateChat(decodedCreatorId, otherUser._id)) {
                 throw new usersAlreadyHaveMutualPrivateChatroomError()
             }
         }
-        
     }
     const expirationDate = await setExpirationDate(isTtl, daysToDie)
     let newChat = new Chat({
