+++ controllers/chatControllers/leaveChatController.js
@@ -2,6 +2,8 @@ const tryCatchWrapper = require('../../middlewares/tryCatchWrapper')
 const User = require('../../models/userModel')
 const Chat = require('../../models/chatroomModel')
 const getCreatorIdFromHeaders = require('../../middlewares/getCreatorIdFromHeaders')
+const { StatusCodes } = require('http-status-codes')
+const { youAreNotInThisChatError } = require('../../errors/chatErrors/chatErrors')
 
 const updaterOptions = {
     new: true,
@@ -17,8 +19,7 @@ const leaveChat = tryCatchWrapper(async (req, res) => {
             user.chats.pull(chatId)
             user.save()
             chat = await Chat.updateOne({ _id: chatId }, { $pull: { users: { user_id: decodedId } } }, updaterOptions)
-            res.status(200).json({
-                success: true,
+            res.status(StatusCodes.OK).json({
                 message: 'Chat left successfully!'
             })
             return
@@ -27,8 +28,7 @@ const leaveChat = tryCatchWrapper(async (req, res) => {
             await Chat.deleteOne({ _id: chatId })
             user.chats.pull(chatId)
             user.save()
-            res.status(200).json({
-                success: true,
+            res.status(StatusCodes.OK).json({
                 message: 'Chat left successfully!',
             })
             return
@@ -38,17 +38,12 @@ const leaveChat = tryCatchWrapper(async (req, res) => {
         chat.save()
         user.chats.pull(chatId)
         user.save()
-        res.status(200).json({
-            success: true,
-            message: 'Chat left successfully!'
+        res.status(StatusCodes.OK).json({
+            message: 'Chat left successfully!',
         })
         return
     }
-    res.status(400).json({
-        success: false,
-        message: 'You are not in this chat!',
-    })
-    return
+    throw new youAreNotInThisChatError()
 })
 
 module.exports = leaveChat
