+++ controllers/forumControllers/getAllForumsController.js
@@ -1,0 +1,20 @@
+const Forum = require('../../models/forumModel')
+const tryCatchWrapper = require('../../middlewares/tryCatchWrapper')
+const { StatusCodes } = require('http-status-codes')
+
+const getAllForums = tryCatchWrapper(async (req, res) => {
+    const page = parseInt(req.query.page) || 0
+    const limit = parseInt(req.query.limit) || 10
+    const skip = page * 10
+    const forums = await Forum.find().skip(skip).limit(limit)
+    if (!forums) {
+        res.status(StatusCodes.NOT_FOUND).json({
+            message: 'No forums found',
+        })
+        return
+    }
+    res.status(StatusCodes.OK).json(forums)
+    return
+})
+
+module.exports = getAllForums
