+++ controllers/forumControllers/getForumByIdController.js
@@ -1,0 +1,17 @@
+const Forum = require('../../models/forumModel')
+const tryCatchWrapper = require('../../middlewares/tryCatchWrapper')
+const { StatusCodes } = require('http-status-codes')
+
+const getForumById = tryCatchWrapper(async (req, res) => {
+    const forum = await Forum.findOne({ '_id.forum_id': req.params.forumId })
+    if (!forum) {
+        res.status(StatusCodes.NOT_FOUND).json({
+            message: 'No forum found',
+        })
+        return
+    }
+    res.status(StatusCodes.OK).json([forum])
+    return
+})
+
+module.exports = getForumById
