+++ routes/router.js
@@ -23,7 +23,7 @@ const protectPath = require('../middlewares/protectPath')
 const checkWetherBannedFromForum = require('../middlewares/checkWetherBannedFromForum')
 const {
     createForum,
-    getAllThreads,
+    getAllThreadsByForumId,
     getAllForums,
     getForumById,
     searchForumByTag,
@@ -31,7 +31,7 @@ const {
     banUserFromForum,
     unbanUserFromForum,
     updateForum,
-    recommendForums
+    recommendForums,
 } = require('../controllers/forumControllers/forumControllers')
 const search = require('../controllers/searchController/searchController')
 const {
@@ -77,7 +77,7 @@ router.route('/register').post(registerUser)
 router.route('/login').post(loginLimiter, loginUser)
 router.route('/verifyEmail/:emailToken').get(verifyEmail)
 
-router.route('/search').post(search)                        //PAGINATION TESTS!
+router.route('/search').post(search)
 
 router.route('/user').put(protectPath, updateUser).delete(protectPath, deleteUser)
 router.route('/user/friends/requests').get(protectPath, getUserRequests)
@@ -88,7 +88,7 @@ router.route('/user/:username').get(getUserProfileByUsername)
 router.route('/user/addHobby').post(protectPath, addHobby)
 
 router.route('/forum').get(getAllForums).post(protectPath, createForum).delete(protectPath, deleteForum)
-router.route('/forum/getAllThreads/:forumId').get(protectPath, checkWetherBannedFromForum, getAllThreads)
+router.route('/forum/getAllThreads/:forumId').get(protectPath, checkWetherBannedFromForum, getAllThreadsByForumId)
 router.route('/forum/getForumsByTag/:tag').get(searchForumByTag)
 router.route('/forum/ban').post(protectPath, banUserFromForum).put(protectPath, unbanUserFromForum)
 router.route('/forum/recommendForums').get(recommendForums)
@@ -113,4 +113,5 @@ router.route('/friends').get(protectPath, getFriends)
 router.route('/friend/:friendName').post(protectPath, makeFriendRequest).delete(protectPath, deleteFriend)
 router.route('/acceptFriendRequest/:requestCreatorName').post(protectPath, acceptFriendRequest)
 router.route('/declineFriendRequest/:requestCreatorName').post(protectPath, declineFriendRequest)
+
 module.exports = router
